info: add di_info_get_device_tag()
Display servers need to be able to identify a monitor regardless of
which port it happens to be connected to. This new API offers the
key to do that.

Add new API to create a string as specified. New API is required because
I want libdisplay-info to reserve the right to change what get_make(),
get_model() and get_serial() may return (e.g. prefer DisplayID blob over
EDID blob) without breaking configuration keys a user may have stored.
When DisplayID becomes supported, the new function wants to maintain
backward compatibility.

For the above reasons, the implementation does not rely on calling
get_make() etc. but depends only on the low-level API.

Signed-off-by: Simon Ser <contact@emersion.fr>
Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>
Co-authored-by: Pekka Paalanen <pekka.paalanen@collabora.com>