cta: do not mess with the YCbCr 4:2:0 cap map
Currently when we find an invalid SVD (bad EDID) in a video block, we
simply avoid adding it to the SVD list. Also, the map between SVD's and
the YCbCr 4:2:0 caps is based on the indices of the SVD's. Combine that
and we have a problem: we mess with the mapping.

This adds a new field "index" to struct di_cta_svd, and we use that to
properly use the mapping from SVD to YCbCr 4:2:0 cap.

Note that this is not a permanent fix that would handle all misformed
EDID's. For instance, this fails if the last VIC in the block is not
valid, or if a video block is skipped. That would require more complex
solutions, but the decision was to not introduce complexity because of
malformed EDID's.

Reported-by: Qinglong Zhong <qingloz@amazon.com>
Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>