cta: try to adjust last Data Block size and parse it
We have multiple Data Blocks followed by Detailed Time Definitions
(DTD). Currently if we encounter a block whose reported size overlaps
DTD, we stop parsing the blocks. Instead, let's assume that this is the
last Data Block and adjust its size to parse it, taking into account the
byte in which DTD starts.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>