cta: Split out HDMI and CTA VIC lookup
The VICs are not tied to the CTA, EDID or DisplayID, but are really just
a mapping between a code and a timing definition.

Given that they are being used in different specs and blocks, it makes
sense to have the VIC definitions and lookups separate and reference
them where needed.

Signed-off-by: Sebastian Wick <sebastian.wick@redhat.com>