cta: do not fail when there's a bad Data Block in a CTA extension
Instead of failing when we fail to parse a Data Block in
parse_data_block(), let's simply skip such block. We want to parse as
much as possible, so skipping a whole CTA extension because of a bad
Data Block is not desirable.

The test EDID added (for Lenovo T24s-28) would fail parsing without this
patch.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>